style: 컨벤션 수정
- 프로퍼티, 메서드 이름 수정
- 줄바꿈 수정
- 주석 추가
- self 클로저 내부 말고는 제외
