+++ OpenMarket/OpenMarket/Extension/UISegmentControl+Extension.swift
@@ -1,14 +1,14 @@
 import UIKit
 
 extension UISegmentedControl {
-    func defaultConfiguration(color: UIColor = UIColor.white) {
+    func setDefaultFontColor(_ color: UIColor = UIColor.white) {
         let defaultAttributes = [
             NSAttributedString.Key.foregroundColor: color
         ]
         setTitleTextAttributes(defaultAttributes, for: .normal)
     }
 
-    func selectedConfiguration(color: UIColor = UIColor.red) {
+    func setSelectedFontColor(_ color: UIColor = UIColor.red) {
         let selectedAttributes = [
             NSAttributedString.Key.foregroundColor: color
         ]
