+++ OpenMarket/OpenMarket/SceneDelegate.swift
@@ -11,7 +11,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
-        let mainViewContoroller = ViewController()
+        let mainViewContoroller = ProductsViewController()
         
         let navigationController = UINavigationController(rootViewController: mainViewContoroller)
